Reimplement the parser using lalrpop 💥
