+++ crates/ditto-cst/Cargo.toml
@@ -8,14 +8,13 @@ license = "BSD-3-Clause"
 doctest = false
 
 [dependencies]
-pest = "2.1"
-pest_derive = "2.1"
 itertools = "0.10"
 serde = { version = "1.0", features = ["derive"] }
 miette = { version = "4.7", features = ["fancy"] }
 thiserror = "1.0"
 lalrpop-util = "0.19.8"
 regex = "1"
+logos = "0.12"
 #simsearch = "xx"   <-- for suggestions
 #unindent = "xx"  <-- might come in useful for smart multi-line strings (like Nix)
 #codespan = "xx" <-- might be a good replacement for our `Span` type
