+++ crates/ditto-cst/golden-tests/parse-errors/bad_package_name.ditto
@@ -1,0 +1,3 @@
+module Test exports (..);
+
+import (foo_bar) Foo.Bar;
