+++ crates/ditto-cst/golden-tests/parse-errors/empty_input.error
@@ -1,7 +1,8 @@
 
-  × syntax error
+  × unexpected eof
    ╭─[golden:1:1]
  1 │ 
    · ▲
-   · ╰── expected: module_keyword
+   · ╰── here
    ╰────
+  help: expected "module"
